# PaperlessREST/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: SWEN3
  description: DMS
  title: Paperless Document Management REST-Server
  version: 1.0.0
servers:
- url: localhost:8081
tags:
- description: Paperless Document Management REST-API
  name: paperless
paths:
  /api/documents:
    get:
      operationId: getDocuments
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/documents"
          description: Success
      summary: Returns a list of documents
      tags:
      - paperless
      x-accepts:
      - application/json
      x-tags:
      - tag: paperless
    post:
      operationId: uploadDocument
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/uploadDocument_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/document"
          description: Document successfully uploaded.
        "400":
          description: Document upload failed due to bad request!
      summary: Upload a new document
      tags:
      - paperless
      x-content-type: multipart/form-data
      x-accepts:
      - application/json
      x-tags:
      - tag: paperless
  /api/documents/{id}:
    delete:
      operationId: deleteDocument
      parameters:
      - description: The id of the document
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Document deleted successfully
        "404":
          description: Document does not exist with this id!
      summary: Delete a document by id
      tags:
      - paperless
      x-accepts:
      - application/json
      x-tags:
      - tag: paperless
    get:
      operationId: getDocument
      parameters:
      - description: The id of the document
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/document"
          description: Success
        "404":
          description: Document does not exist with this id!
      summary: Retrieve a document by id
      tags:
      - paperless
      x-accepts:
      - application/json
      x-tags:
      - tag: paperless
    put:
      operationId: updateDocument
      parameters:
      - description: The id of the document
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/document"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/document"
          description: Success
        "404":
          description: Document does not exist with this id!
      summary: Update a document by id
      tags:
      - paperless
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: paperless
  /api/documents/search:
    get:
      operationId: searchDocuments
      parameters:
      - description: The search query to match against document content
        explode: true
        in: query
        name: query
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/documents"
          description: Success
        "400":
          description: Invalid search query
      summary: Search for documents by content
      tags:
      - paperless
      x-accepts:
      - application/json
      x-tags:
      - tag: paperless
components:
  schemas:
    document:
      example:
        summary: summary
        uploadDate: 2000-01-23T04:56:07.000+00:00
        fileSize: 6
        author: author
        id: 0
        mimeType: mimeType
        title: title
        content: content
        originalFilename: originalFilename
      properties:
        id:
          description: The unique id of the document
          format: int64
          type: integer
        title:
          description: The title of the document
          type: string
        author:
          description: The author of the document
          type: string
        content:
          description: The content of the document
          type: string
        summary:
          description: AI-generated summary of the document
          type: string
        originalFilename:
          description: The original filename of the uploaded document
          type: string
        mimeType:
          description: "The MIME type of the document (e.g., application/pdf)"
          type: string
        fileSize:
          description: The size of the file in bytes
          format: int64
          type: integer
        uploadDate:
          description: When the document was uploaded
          format: date-time
          type: string
      required:
      - title
    documents:
      items:
        $ref: "#/components/schemas/document"
      type: array
    uploadDocument_request:
      properties:
        file:
          description: "The document file (PDF, DOCX, TXT, etc.)"
          format: binary
          type: string
        title:
          description: Optional document title
          type: string
        author:
          description: Optional document author
          type: string
      required:
      - file